requirements.txt
Django>=5.2.10,<6.0

gunicorn>=22.1.0

whitenoise>=6.5.0

dj-database-url>=1.0.0

python-dotenv>=1.0.0

psycopg2-binary>=2.9.10

Pillow>=10.0.0

requests>=2.31.0

supabase>=2.0.0
razorpay>=3.0.0
cloudinary>=1.40.0
django-cloudinary-storage>=0.3.0
